# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Aegis Panel API
  version: 0.7.0
  description: |
    Aegis — self-hosted VPN control panel. Multi-core (sing-box / Xray /
    Hysteria 2), BYO Node, Cascade Topology, MCP-driven, decoy sites,
    port masquerading.

    v0.7.0 surface (this document):
      - /auth/*  (login, refresh, me, change-password)
      - /nodes/*  (CRUD — admin)
      - /nodes/{nodeId}/inbounds/*  (CRUD — admin)
      - /hosts/*  (CRUD — admin)
      - /plans/*  (CRUD — admin; v0.6.0)
      - /users/*  (CRUD + rotate-token — admin)
      - /panelcfg/*  (sub_path rotation — admin)
      - /sub/{token}  (public subscription render — anonymous)
      - /audits/*  (audit log read — admin)
      - /backups/*  (CRUD + restore — admin; v0.5.0)
      - /webhooks/*  (CRUD + deliveries + DLQ + test — admin; v0.7.0)
      - /health  (liveness — anonymous)
      - /cores  (provider catalog — admin)

    Out of scope for this spec (added in subsequent v0.2.x PRs):
      - /admin/*  (admin principal CRUD — PR-J)
      - /cabinet/*  (end-user cabinet)

  license:
    name: AGPL-3.0-or-later
  contact:
    name: Aegis Panel
    url: https://github.com/QAdversif/AegisPanel
servers:
  - url: https://panel.example.com/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - BearerAuth: []
paths:
...
  /nodes:
    get:
      tags: [nodes]
      summary: List nodes
      description: |
        Every VPN node registered with the panel. The list is sorted
        by `name` ascending. v0.1.0 does not paginate — the panel
        scale is "tens of nodes" so the full list is fine.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeListResponse'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [nodes]
      summary: Register a node
      description: |
        Register a new VPN node with the panel. The agent
        registration handshake is a separate step that lands with
        the BYO Node flow in v0.3 (per the v9 §3.3 roadmap); for now
        the panel stores the node metadata and the operator SSHes
        into the node by hand to install the agent.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Name or address already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /nodes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [nodes]
      summary: Get a node
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [nodes]
      summary: Update a node
      description: |
        Partial update (the same PUT semantics the rest of the
        v0.2.0 CRUD surfaces use). Absent keys mean "leave alone".
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [nodes]
      summary: Delete a node
      responses:
        '204':
          description: No content
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /nodes/{id}/provision:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [nodes]
      summary: Provision a node (v0.3.0 BYO Node flow)
      description: |
        Runs the bootstrap workflow: SSH probe with
        the supplied key, install the placeholder agent
        systemd unit, and transition the node to either
        `online` (success) or `offline` (failure). Only
        nodes in a provisionable state (`new` or
        `offline`) accept the call; a node already
        `online` returns 409.

        v0.3.0 ships the synchronous `Provision` method.
        v0.5.0 adds a kick-off+poll mode for large
        fleets (returns 202 Accepted with a job id).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeProvisionRequest'
      responses:
        '200':
          description: Provision completed (state is now online or offline)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeProvisionResponse'
        '400':
          description: Validation error (missing ssh_private_key, unknown tofu_policy, malformed body)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Node not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Node is not in a provisionable state (e.g. already online)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '502':
          description: Provision failed (SSH unreachable, install error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `Authorization: Bearer <jwt-access-token>`. Tokens are
        HS256-signed by Aegis (`iss=aegis-panel`,
        `aud=aegis-panel/v1`).
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: |
            Human-readable error message. For security reasons
            the auth endpoints collapse several failure modes
            into a single generic message; do not branch on
            the message text.
...
    NodeState:
      type: string
      enum: [new, online, draining, offline, disabled]
    TofuPolicy:
      type: string
      enum: [reject, accept-and-append]
      description: |
        Trust-on-first-use policy for the SSH host key.

        - `reject` (default): the operator must paste the
          expected SHA256 fingerprint in `expected_fingerprint`
          before the panel will connect. Safe mode.
        - `accept-and-append`: on first contact, accept the
          server's key, append it to the panel's known_hosts,
          and continue. The fingerprint is returned in
          the response so the operator can verify it after.
    Node:
      type: object
      required: [id, name, region, state, address, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 64
        region:
          type: string
          maxLength: 64
        state:
          $ref: '#/components/schemas/NodeState'
        address:
          type: string
          description: 'SSH endpoint (host:port) the panel uses to reach the node.'
        capacityHint:
          type: string
          maxLength: 64
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    NodeCreateRequest:
      type: object
      required: [name, region, address]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        region:
          type: string
          minLength: 1
          maxLength: 64
        address:
          type: string
          minLength: 1
        capacityHint:
          type: string
          maxLength: 64
    NodeUpdateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        region:
          type: string
          minLength: 1
          maxLength: 64
        address:
          type: string
        state:
          $ref: '#/components/schemas/NodeState'
        capacityHint:
          type: string
          maxLength: 64
        tags:
          type: array
          items:
            type: string
    NodeProvisionRequest:
      type: object
      description: |
        Exactly one of `ssh_private_key` or `ssh_password`
        must be set. The `ssh_password` path is the
        first-time-install UX: the operator pastes the
        VPS root password, the panel SSHes in to install
        the agent, and the agent switches to bearer-
        token auth so the password is never reused.
      properties:
        ssh_port:
          type: integer
          minimum: 1
          maximum: 65535
          description: Per-call override. Zero/omitted means "use the service-wide default (22)".
        ssh_user:
          type: string
          description: Per-call override. Empty/omitted means "use the service-wide default (root)".
        ssh_private_key:
          type: string
          format: password
          description: |
            Operator-pasted private key (PEM, no passphrase).
            The panel does NOT store this; the install is
            the only consumer. Mutually exclusive with
            `ssh_password`.
        ssh_password:
          type: string
          format: password
          description: |
            Operator's SSH login password for first-time
            auth on a fresh node. The panel does NOT
            store this; the install is the only consumer.
            The agent switches to bearer-token auth so
            the password is never reused. Mutually
            exclusive with `ssh_private_key`.
        tofu_policy:
          $ref: '#/components/schemas/TofuPolicy'
        expected_fingerprint:
          type: string
          description: |
            Operator-confirmed SHA256 fingerprint. Required
            when `tofu_policy` is `reject`; ignored when
            `accept-and-append`. Format: `SHA256:base64`.
    NodeProvisionResponse:
      type: object
      required: [node_id, new_state]
      properties:
        node_id:
          type: string
          format: uuid
        new_state:
          $ref: '#/components/schemas/NodeState'
        install_stage:
          type: string
          description: |
            Best-effort stage tag from the provisioner
            (`connect` / `upload` / `env` / `unit` /
            `verify`). Surface in the UI tooltip for the
            "retry" button.
        install_error:
          type: string
          description: Set when `new_state` is `offline`. Empty string on success.
        verify_latency:
          type: string
          description: ISO-8601 duration string (e.g. `PT2.5S`) for the systemd is-active poll.
    NodeListResponse:
      type: object
      required: [nodes]
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/Node'
...